% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{ame}
\alias{ame}
\title{Average Marginal Effects}
\usage{
ame(model, cont_vars, fac_vars, nsim = 100, seed = 1)
}
\arguments{
\item{model}{an object of class \code{glm}.}

\item{cont_vars}{a character vector containing the names of continuous model variables.}

\item{fac_vars}{a character vector with factor model variables.}

\item{nsim}{number of simulations. Default is \code{100}.}

\item{seed}{Seed value for the random number generator. Default value is \code{1}.}
}
\value{
\code{ame} returns an object of class \code{"ame"}. The functions \code{summary} (\code{summary.ame}) and \code{print} (\code{print.summary.ame}) are used to display a summary of the computed average marginal effects.

An object of class \code{"ame"} is a list that contains at least the following components:

\item{\code{continuous}}{a named list containing average marginal effects for each continuous variable. If no variable is specified, defaults to \code{NULL}. The list has attributes \code{variable_type} and \code{variable_names}.}
\item{factor}{a named list containing first differences for each factor variable. If no factor is specified, defaults to \code{NULL}. The list has attributes \code{variable_type} and \code{variable_names}.}

Furthermore, objects of class \code{"ame"} store the model family and link in the attributes \code{model_family}, \code{model_link}, and the random number generator seed and number of draws from the multivariate normal in the attributes \code{seed} and \code{n_sim}.
}
\description{
Compute average marginal effects for generalized linear and generalized additive models using Krinsky-Robb Monte-Carlo simulation.
}
\examples{
# generate some data (factor, dummy, continuous)
set.seed(100)

y <- rbinom(100, size = 1, prob = 0.5)
mcat <- as.factor(rpois(100, 3))
dummy <- as.factor(rbinom(100, size = 1, prob = 0.5))
cont <- runif(100, -10, 10)
data <- data.frame(y, mcat, dummy, cont)

# run glm model
glm_fit <- glm(y ~ mcat + dummy * cont, data = data, family = binomial(link = "logit"))

summary(glm_fit)

# compute average marginal effects
glm_ame <- ame(glm_fit, cont_vars = c("cont"), fac_vars = c("mcat", "dummy"), nsim = 1000)

summary(glm_ame)

# linear model using glm
lm1 <- glm(mpg ~ cyl + hp + wt, data = mtcars, family = gaussian)
lm2 <- glm(mpg ~ cyl * hp + wt, data = mtcars, family = gaussian)

summary(lm1)
summary(lm2)

# compute average marginal effects
lm1_ame <- ame(lm1, cont_vars = c("cyl", "hp", "wt"), nsim = 1000, seed = 99)
lm2_ame <- ame(lm2, cont_vars = c("cyl", "hp", "wt"), nsim = 1000, seed = 99)

# display results
summary(lm1_ame)
summary(lm2_ame)

}
\author{
Julia Partheymueller, Alexander Staudt (\email{astaudt@mail.uni-mannheim.de})
}
\references{
Krinsky I, Robb AL (1991). "Three Methods for Calculating the Statistical Properties of Elasticities: A Comparison." \emph{Empirical Economics} 16(2), 199--209.

Brambor T, Clark WR, Golder M (2006). "Understanding Interaction Models: Improving Empirical Analyses." \emph{Political Analysis} 14(1), 63--82. url: \url{http://localgov.fsu.edu/readings_papers/Research Methods/Brambor_et_al_Multipolicative_Interactions.pdf}
}
\seealso{
King G, Tomz M, Wittenberg J (2000). "Making the Most of Statistical Analyses: Improving Interpretation and Presentation." \emph{American Journal of Political Science} 44(2), 347--361. url: \url{http://www.polmeth.wustl.edu/files/polmeth/king98f.pdf}.
}

