% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAME.R
\name{compAME}
\alias{compAME}
\title{Compute marginal average effects}
\usage{
compAME(object, ...)
}
\arguments{
\item{object}{Model object}

\item{x}{Feature}

\item{delta}{Delta for computing AME}
}
\value{
AME
}
\description{
Compute marginal average effects
}
\examples{
4

}
