% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAME.R
\name{compAME.WrappedModel}
\alias{compAME.WrappedModel}
\title{WrappedModel class from mlr package}
\usage{
\method{compAME}{WrappedModel}(object, task, features, predict.fun = NULL,
  delta = NULL, parallel = FALSE, ...)
}
\description{
WrappedModel class from mlr package
}
