% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{margex}
\alias{margex}
\title{Stata's margex data set.}
\source{
\url{http://www.stata-press.com/data/r13/r.html}.
}
\description{
Stata's margex data set, which is fictitious data with a dichotomous outcome variable and various demographics.
}
\references{
\url{http://www.stata.com/manuals13/rmargins.pdf}
}
\keyword{data}
