% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionDerivative.R
\name{partitionDerivative}
\alias{partitionDerivative}
\title{Generate partitions with "similar" derivative}
\usage{
partitionDerivative(x, y, max.splits)
}
\arguments{
\item{max.splits}{}
}
\value{
Split points
}
\description{
Generate partitions with "similar" derivative
}
