% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAME.R
\name{computeAME}
\alias{computeAME}
\title{Compute Average Marginal Effects}
\usage{
computeAME(model, data, features, ...)
}
\arguments{
\item{model}{[\code{\link[mlr]{WrappedModel}} | \code{any}]\cr
A model object. Can be also an object of class \code{\link[mlr]{WrappedModel}}.}

\item{data}{[\code{logical(1)}]\cr
The data set that was used to fit the model.}

\item{features}{[\code{logical(1)}]\cr
The features for which the average marginal effects should be computed.}

\item{...}{Further options passed down to the \code{\link[numDeriv]{grad}} function.}
}
\description{
Computes the average marginal effects for specified features.
}
